Refactor mock sessions and mock queries to reduce complexity.
